requirements.txt
Flask>=3.0,<4
google-genai>=1.65.0,<2
python-dotenv>=1.0.1,<2
colorama>=0.4.6,<1
gunicorn>=22.0.0,<23; platform_system != "Windows"
waitress>=3.0.0,<4; platform_system == "Windows"
